Wrote copy and paste
I guess I had to write the code for those keymappings. I'm not sure why
they weren't handled like the other key bindings.  My suspicion is that
textarea handles cut and copying to the clipboard without any additional
code for them and it's possible that cut and copy are somehow disabled
either because of either how Ace is set up right or because that's how
textareas are in Windows.