Work towards restoring the Recent files header
I think I stumbled upon what was wrong with the recent files list.  My
belief is that it was a memory leak after all.  The information for each
file is loaded through anonymous functions executed asynchronously, so a
lot of data held in the parent's scope gets carried into these functions
"activation record" (I'm not sure if that's what it's called in
Javascript).  I've mastered scheduling and timing of Windows Promises,
which was the major change in 1.1.1.  I cleaned up some of the variable
declarations and references to solve the leak.  In the future, I need to
be more mindful about what data is referenced from within a Promise.